Factor out schema_base.py.
